Add explicity package name to tox py.test command
If there is, for example, a virtualenv setup inside the project's root
directory, running tox will result in errors. This happens due to
py.test trying to execute the tests for the packages inside the
virtualenv.

By explicitly naming the package on which tests should be run, this
problem is mitigated.