Fix wrongly escaped links in admin view lists
The `allow_tags` attribute is deprecated in Django 2.0:
https://django.readthedocs.io/en/2.0.x/internals/deprecation.html

As a consequence, "Update" edit links are no longer clickable for admin
models that inherit from ConfigurationModelAdmin.

This will close CRI-189 once a new release of django-config-models is
produced and the dependency is upgraded in edx-platform.

We take the opportunity to bump the version number to v2.0.2.