chore: remove tox-travis dependency
tox-travis is a project that is no longer maintained: https://github.com/tox-dev/tox-travis
It was necessary in this repo to be used from within .travis.yml: https://github.com/openedx/django-config-models/pull/35
We no longer use this file, so we can get rid of the dependency. This
allows us to also upgrade tox to v4, which in turn updates quite a few
dependencies.

Note that in tox.ini, "whitelist_externals" was renamed to
"allowlist_externals" in tox 3.18.0 https://tox.wiki/en/3.18.0/config.html#conf-allowlist_externals